\name{Config.Version}
\alias{Config.Version}
\title{
SciServer release tag
}
\description{
Defines the SciServer release version tag (string), to which this package belongs.
}
\usage{Config.Version}
\author{
Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>
}
\seealso{
\code{\link[SciServer]{Config.isSciServerComputeEnvironment}},\code{\link[SciServer]{Config.CasJobsRESTUri}},\code{\link[SciServer]{Config.AuthenticationURL}},\code{\link[SciServer]{Config.SciDriveHost}},\code{\link[SciServer]{Config.SkyQueryUrl}},\code{\link[SciServer]{Config.DataRelease}},\code{\link[SciServer]{Config.SkyServerWSurl}},\code{\link[SciServer]{Config.Version}},\code{\link[SciServer]{Config.KeystoneTokenPath}}
}
\references{
\href{http://www.sciserver.org}{http://www.sciserver.org}\cr
\href{http://skyserver.sdss.org}{http://skyserver.sdss.org}\cr
\href{http://www.github.com/sciserver/SciScript-R}{http://www.github.com/sciserver/SciScript-R}
}
\keyword{SciServer}
