\name{SciScript-package}
\alias{SciScript-package}
\alias{SciScript}
\docType{package}
\title{
SciScript
}
\description{
SciScript
}
\details{
\tabular{ll}{
Package: \tab SciScript\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-22\cr
License: \tab GPL\cr
}

}
\author{
Manuchehr Taghizadeh-Popp, Gerard Lemson
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>
}
\references{
No references for now
}
\keyword{ package }
\seealso{
}
\examples{
dt <- runSqlSearch("select 1")
}
