\name{SciServer}
\alias{SciServer}
\docType{package}
\title{
Access to SciServer tools.
}
\description{
This package gives access to Sciserver tools.
}

\details{

\tabular{ll}{
Package: \tab SciServer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-22\cr
}}
\author{
Manuchehr Taghizadeh-Popp, Gerard Lemson \cr
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>
}
\references{
\href{http://www.sciserver.org}{http://www.sciserver.org}\cr
\href{http://www.github.com/sciserver/SciScript-R}{http://www.github.com/sciserver/SciScript-R}
}
\keyword{ package }
\seealso{ }
